feat(content): add title and tagline to front page
