+++ content/pages/home/header.json
@@ -1,0 +1,4 @@
+{
+  "title": "Goal",
+  "content": "The Universitat de Lleida MVG Portal aims to accelerate the progress of research and science, increase the quality of data, and foster collaboration between research and educational institutions. On the research portal, data scientists and researchers can find, consume and share research data sets and algorithms. Depending on the publisher’s decision, the data assets can be consumed either by granting compute or download access to the data assets."
+}
