+++ src/components/pages/Home/Header.tsx
@@ -1,32 +1,17 @@
 import React, { ReactElement } from 'react'
 import styles from './Header.module.css'
-import { graphql, useStaticQuery } from 'gatsby'
 import Markdown from '../../atoms/Markdown'
-import Img from 'gatsby-image'
+import Img, { FluidObject } from 'gatsby-image'
 
-const contentQuery = graphql`
-  query TaglineQuery {
-    file(relativePath: { eq: "pages/home/intro.json" }) {
-      childHomeJson {
-        title
-        content
-      }
-    }
-    image: file(relativePath: { eq: "microscope.png" }) {
-      childImageSharp {
-        fluid {
-          ...GatsbyImageSharpFluid
-        }
-      }
-    }
-  }
-`
-
-export default function Header(): ReactElement {
-  const data = useStaticQuery(contentQuery)
-  const { title, content } = data.file.childHomeJson
-  const image = data.image.childImageSharp.fluid
-
+export default function Header({
+  title,
+  content,
+  image
+}: {
+  title: string
+  content: string
+  image: FluidObject
+}): ReactElement {
   return (
     <div className={styles.container}>
       <div className={styles.content}>
