+++ src/components/pages/Home/Intro.module.css
@@ -1,0 +1,18 @@
+.powered {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin-bottom: var(--spacer);
+}
+
+.powered h1 {
+  text-align: center;
+  font-weight: var(--font-weight-bold);
+}
+
+.tagline {
+  margin-bottom: var(--spacer);
+  max-width: 40rem;
+  text-align: center;
+  font-size: var(--font-size-large);
+}
