+++ src/components/pages/Home/index.module.css
@@ -1,9 +1,3 @@
-.powered {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
 .topicsWrapper {
   display: flex;
   justify-content: center;
